Moved records updater to a proper namespace
